+++ Assets/Scripts/BINGO.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a2fdb018cc7ddea43a4304ec9a8ce74f
+guid: a68c5029f6394d44cbabbc39c1161af0
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
