+++ Assets/Scripts/BINGO/AreaInfo.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+using TMPro;
+
+public class AreaInfoView : MonoBehaviour
+{
+    [SerializeField] private UnityEngine.UI.Image currentImage;
+
+    public UnityEngine.UI.Image CurrentImage => currentImage;
+}
