+++ Assets/Scripts/BINGO/AreaInfo.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d5a1ffced5ffa66459d869b365c5cdf6
+guid: 428cc2b97515eee41b405cac77988d39
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
