+++ Assets/Scripts/BINGO/BingoManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f8fe27e64274d5d4aaf6079942d27e14
+guid: 706036133b3eb564d82efd9cf35e666b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
