+++ Assets/Scripts/BINGO/CardAreaView.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 76bc1c94a7fbd5a4b860efdbde0fd88b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
