+++ Assets/Scripts/BINGO/CardView.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 46a3eb1ac613c414abd57f8275985072
+guid: 05749fbb9b916574bafe1aa1c5ead665
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
